+++ es6/xml-templater.js
@@ -83,11 +83,7 @@ module.exports = class XmlTemplater {
 		this.setModules({
 			inspect: { filePath: this.filePath, xmllexed: this.xmllexed },
 		});
-		const { lexed, errors: lexerErrors } = Lexer.parse(
-			this.xmllexed,
-			this.delimiters,
-			this.syntax
-		);
+		const { lexed, errors: lexerErrors } = Lexer.parse(this);
 		this.allErrors = this.allErrors.concat(lexerErrors);
 		this.lexed = lexed;
 		this.setModules({