+++ preversion.bash
@@ -12,5 +12,5 @@ npm run test:typings
 npm run verifypublishsize
 npm run test:browser
 cp es6/expressions.js es6/expressions-ie11.js js/text.js .
-sed -E -i 's@require...@require(\"./js@g' text.js
+sed -E -i 's@require\("\.@require(\"./js@g' text.js
 node es6/test-text.js